backend/requirements.txt
# FELICITA - Dependencias Python Principal
# Este archivo incluye todas las dependencias necesarias para desarrollo local

# ===========================================
# DJANGO CORE
# ===========================================
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0

# ===========================================
# BASE DE DATOS MYSQL
# ===========================================
mysqlclient==2.2.4

# ===========================================
# CACHE Y REDIS
# ===========================================
redis==5.0.1
django-redis==5.4.0

# ===========================================
# VARIABLES DE ENTORNO
# ===========================================
python-decouple==3.8

# ===========================================
# DOCUMENTACIÓN API
# ===========================================
drf-spectacular==0.27.1

# ===========================================
# UTILIDADES WEB
# ===========================================
requests==2.31.0
urllib3==2.1.0

# ===========================================
# PROCESAMIENTO DE DATOS
# ===========================================
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9

# ===========================================
# FECHAS Y TIEMPO
# ===========================================
python-dateutil==2.8.2
pytz==2023.3.post1

# ===========================================
# VALIDACIONES Y FORMATOS
# ===========================================
validators==0.22.0
python-stdnum==1.19

# ===========================================
# ARCHIVOS Y MEDIA
# ===========================================
Pillow==10.2.0
python-magic==0.4.27

# ===========================================
# LOGGING Y MONITORING
# ===========================================
python-logging-loki==0.3.1
structlog==23.2.0

# ===========================================
# CRIPTOGRAFÍA Y SEGURIDAD
# ===========================================
cryptography==41.0.7
bcrypt==4.1.2

# ===========================================
# TESTING
# ===========================================
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# ===========================================
# DESARROLLO
# ===========================================
django-debug-toolbar==4.2.0
ipython==8.18.1
black==23.12.1
flake8==7.0.0
isort==5.13.2

# ===========================================
# REPORTES Y PDF
# ===========================================
reportlab==4.0.8
weasyprint==60.2

# ===========================================
# INTEGRACIÓN ESPECÍFICA PERÚ
# ===========================================
# Para validaciones RUC/DNI peruanos
#python-peru==0.1.0

# ===========================================
# SERIALIZACIÓN
# ===========================================
djangorestframework-xml==2.0.0
djangorestframework-csv==3.0.2

# ===========================================
# BACKGROUND TASKS
# ===========================================
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0

# ===========================================
# MANEJO DE ARCHIVOS
# ===========================================
django-storages==1.14.2
boto3==1.34.13

# ===========================================
# OPTIMIZACIÓN
# ===========================================
django-cachalot==2.6.1
django-compressor==4.4

# ===========================================
# INTERNACIONALIZACIÓN
# ===========================================
django-rosetta==0.9.9

whitenoise==6.6.0